Net5 (#4)
major cleanup for upgrade to .net 5

+++ Arnible.MathModeling/Geometry/ICoordinate.cs
@@ -2,7 +2,7 @@
 {
   interface ICoordinate<T> where T: struct
   {
-    uint DimensionsCount { get; }
+    ushort DimensionsCount { get; }
 
     T AddDimension();
   }
